TUC-147 Adding percentiles output in CLI
The percentiles are written in a single file.

In test/scripts/octave the file display_percentiles.m contains a function
to generate 3 figures (population, apriori, aposteriori) if the requests
respect a certain convention.

It works well with the query file test/data/queries/test_imatinib_percentiles.xml

 Changes to be committed:
	modified:   src/tucucommon/xmldocument.cpp
	modified:   src/tucucore/computingresponseexport.cpp
	modified:   src/tucuquery/queryimport.cpp
	modified:   src/tucuquery/querytocoreextractor.cpp
	modified:   src/tucuquery/querytocoreextractor.h
	modified:   src/tucuquery/requestdata.cpp
	modified:   src/tucuquery/requestdata.h
	new file:   test/data/queries/test_imatinib_percentiles.xml
		* Query file for testing percentiles on imatinib
	modified:   test/data/queries/xml_query.xsd
		* percentiles ranks are double instead of int
	new file:   test/scripts/octave/display_percentiles.m
		* An octave script to display percentiles

+++ src/tucuquery/queryimport.cpp
@@ -774,12 +774,12 @@ unique_ptr<RequestData> QueryImport::createRequest(Tucuxi::Common::XmlNodeIterat
 
     Common::XmlNodeIterator percentilesRootIterator = _requestRootIterator->getChildren(PERCENTILES_NODE_NAME);
     Common::XmlNodeIterator percentilesIterator = percentilesRootIterator->getChildren();
-    vector<unsigned short> percentiles;
+    vector<double> percentiles;
     while(percentilesIterator != percentilesIterator.none()) {
         string percentileValue = percentilesIterator->getValue();
-        unsigned short percentile = 0;
+        double percentile = 0;
         try {
-            percentile = static_cast<unsigned short>(stoul(percentileValue));
+            percentile = static_cast<double>(std::stod(percentileValue));
         } catch (invalid_argument e) {
             percentile = 0;
         } catch (out_of_range e) {
